feat: system-consolidation-validation-test-file-exemption-burn-down-004 - Preserve orchestrator compatibility and work-propagation evidence
